Refactorize main.rb file
